requirements.txt
Flask==2.1.4
Flask-RESTful==0.3.9
PyJWT==2.3.0
pymongo==4.1.1
Flask-Cors==3.0.10